Add files via upload
In this project, I have extracted features which are the mean and standard deviation of the images of size 28*28 for Fashion-MNIST data and estimated parameters for these features. Further, I performed Naïve Bayes Classification with the help of the parameters we estimated earlier. Also, I have classified the data with a Logistic Regression model using gradient ascent. I also computed the accuracy for both the models and concluded that the Logistic Regression model performed better than Naïve Bayes Classifier for this use case.